Code refactor/minimizing, stars component
