+++ src/assets/style/contact.css
@@ -14,6 +14,7 @@
     justify-content: center;
     align-items: center;
     max-width: 100vw;
+    height: 81vh;
     margin-bottom: 0;
     bottom: 0px;
     background: rgb(0,0,0);
@@ -22,7 +23,9 @@
 .contact-right-left-side-container {
     display: flex;
     width: 100vw;
+    
     flex-direction: row;
+    justify-content: flex-end;
     align-items: flex-end;
 }
 .contact-left-side {
@@ -49,7 +52,6 @@
 }
 
 .contact-right-side {
-    position: relative;
     display: flex;
     flex-wrap: wrap;
     background-color: rgba(255, 255, 255, 0.075);
