+++ src/assets/style/home.css
@@ -1,31 +1,25 @@
 .home-container {
     display: flex;
     position: relative;
-    flex-direction: column;
+    flex-direction: row;
     justify-content: center;
     align-items: center;
+
     max-width: 100vw;
-    height: 81vh;
-    min-height: 750px;
-    margin-bottom: 0;
-    top: 0px;
-    /* overflow: hidden; */
+    height: 100%;
+
     background: rgb(0,0,0);
     background: linear-gradient(70deg, rgba(0,0,0,0) 65%, var(--purple-dark) 100%);
 }
-.home-right-left-side-container {
-    display: flex;
-    width: 100vw;
-    flex-direction: row;
-}
+
 .home-hub {
     position: absolute;
     display: flex;
     flex-direction: row;
     justify-content: center;
     align-items: center;
-    top: 170px;
     z-index: 5;
+    top: 20%;
 }
 
 .home-picture {
@@ -124,8 +118,11 @@
     transform: scale(1.05);
 }
 
-.home-left-side {
+.home-background {
     display: flex;
+    position: absolute;
+    bottom: 0;
+    left: 0;
     flex-direction: column;
     width: 50vw;
     height: 71vh;
@@ -136,7 +133,7 @@
     
 }
 
-.home-left-side img {
+.home-background img {
     image-rendering: pixelated;
     position: absolute;
     width: 50vw;
