+++ src/assets/style/portfolio.css
@@ -5,87 +5,28 @@
     display: none;
 }
 
-.portfolio-container {
-    display: flex;
-    position: relative;
-    flex-direction: column;
-    justify-content: center;
-    align-items: center;
-    /* overflow: hidden; */
-    max-width: 100vw;
-    /* min-height: 81vh; */
-    margin-bottom: 0;
-    bottom: 0px;
-    background: rgb(0,0,0);
-    background: linear-gradient(70deg, rgba(0,0,0,0) 65%, var(--purple-dark) 100%);
-}
-.portfolio-right-left-side-container {
-    display: flex;
-    width: 100vw;
-    flex-direction: row;
-    align-items: flex-end;
-}
-.portfolio-left-side {
-    display: flex;
-    flex-direction: column;
-    width: 30vw;
-    height: 81vh;
-    justify-content: center;
-    z-index: 0;
-}
-
-.portfolio-left-side img {
-    image-rendering: pixelated;
-    position: absolute;
-    width: 50vw;
-    left: -20vw;
-    background-repeat: no-repeat;
-}
-.portfolio-trees, .portfolio-city1, .portfolio-city2 {
-    bottom: 0px;
-}
-.portfolio-mountains {
-    bottom: 30px;
-}
-
-.portfolio-right-side {
-    position: relative;
-    display: flex;
-    flex-wrap: wrap;
-    background-color: rgba(255, 255, 255, 0.075);
-    flex-direction: row;
-    top: 0;
-    width: 68vw;   
-    min-height: 49.5vh;
-    align-items: center;
-    justify-content: center;
-    margin-left: 1vw;
-    margin-bottom: 1vw;
-    border-bottom-left-radius: 5px;
-    border-bottom-right-radius: 5px;
-}
-.portfolio-right-side::after {
-    content:"";
-    position: absolute;
-    background-color: rgba(255, 255, 255, 0.075);
-    height: 10vh;
-    width: 37.8vw;
-    right: 0;
-    bottom: 49.5vh;
-    border-top-right-radius: 5px;
-}
 .portfolio-title {
     display: flex;
     text-align: center;
+    justify-content: center;
     align-items: center;
     position: absolute;
     z-index: 5;
-    width: 60vw;
-    bottom: 49.5vh;
-    margin-bottom: .5vw;
-    margin-left: -.5vw;
-    left:2.5vw;
-    border-top-right-radius: 20px;
+    width: 100%;
+    height: 8.5vw;
+    bottom: calc(49.5vh + .5vw);
+}
+.portfolio-title::after {
+    content:"";
+    display: flex;
+    position: absolute;
+    right: 0;
+    bottom: 0;
+    background-color: rgba(255, 255, 255, 0.075);
+    height: 10vh;
+    width: calc(100% - 30.5vw);
+    margin-bottom: -.5vw;
+    border-top-right-radius: 5px;
 }
 .portfolio-folder-icon-container {
     position: absolute;
@@ -93,7 +34,7 @@
     justify-content: center;
     align-items: center;
     background-color: #26243d;
-    left: -9.75vw;
+    left: -7.9vw;
     border-radius: 250px;
     width: 14.5vw;
     height: 14.5vw;
@@ -108,31 +49,37 @@
     transform: scale(0.975);
     filter: drop-shadow(0px 0px 10px var(--blue))
             drop-shadow(0px 0px 4px var(--blue));
-    left: 0;
+    left: -2%;
 }
 .portfolio-title h1 {
+    position: absolute;
+    left: 0;
+    display: flex;
+    justify-content: flex-end;
+    align-items: center;
     text-transform: lowercase;
     font-size: 7vw;
     letter-spacing: .07em;
-    top: 0px;
-    padding: .6vw 1.5vw 1.6vw 5vw;
+    width: 30vw;
+    height: 100%;
+    padding-bottom: .5vw;
+    padding-right: 1.5vw;
     color: var(--blue-highlighted);
     text-shadow: 0px 0px 7px var(--blue),
     0px 0px 12px var(--blue);
-    filter: drop-shadow(0px 0px 12px #7fdfff7c);
     z-index: 2;
     line-height: 1;
     margin: 0;
-    border-top-right-radius: 25px;
+    border-top-right-radius: 35px;
     background-color: #26243d;
-    filter: drop-shadow(2px 4px 5px rgba(0, 0, 0, 0.322));
+    filter: drop-shadow(2px 4px 4px rgba(0, 0, 0, 0.322));
 }
 .portfolio-title h1::before {
     content:"";
     position: absolute;
-    left: 5.25vw;
-    bottom: 2.1vw;
-    width: 72%;
+    right: 2.2vw;
+    bottom: 1.3vw;
+    width: 67%;
     border-radius: 20px;
     padding-right: 10px;
     height: 2px;
@@ -143,7 +90,7 @@
 .project-tabs {
     position: absolute;
     display: flex;
-    left: 32vw;
+    left: 32.5vw;
     bottom: 59.5vh;
     height: 50px;
     font-size: 2em;
