+++ src/components/Navigation.jsx
@@ -1,8 +1,5 @@
 import React from 'react';
 import { useState } from 'react';
-import stars1 from '../assets/stars1.png';
-import stars2 from '../assets/stars2.png';
-import stars3 from '../assets/stars3.png';
 
 
 function Navbar({ currentPage, handlePageChange }) {
@@ -11,10 +8,6 @@ function Navbar({ currentPage, handlePageChange }) {
 
     return (
         <>
-            {/* TODO: Remove these from navbar and add to wrapper */}
-            <img src={stars1} alt="stars1" id="stars1" className="object" data-value="6" />
-            <img src={stars2} alt="stars2" id="stars2" className="object" data-value="3" />
-            <img src={stars3} alt="stars3" id="stars3" className="object" data-value="2" />
             <nav className="navbar">
                 <div className="navbar-bg"></div>
                 <a href="#home" className="navbar-brand" onClick={() => handlePageChange("Home")}>
@@ -31,7 +24,6 @@ function Navbar({ currentPage, handlePageChange }) {
                     <div className="nav">
                         <a href="#home" id="home" 
                             onClick={() => {
-                                let navbar = document.querySelector('.navbar-collapse');
                                 handlePageChange("Home");
                             }}
                             className={currentPage === "Home" ? "nav-link nav-link-current" : "nav-link"}>
